refactor: ui 패키지 각 도메인 하위로 이동
