+++ src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.menu.dto;
 
 import java.math.BigDecimal;
 import java.util.List;
