+++ src/main/java/kitchenpos/menu/dto/MenuResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.response;
+package kitchenpos.menu.dto;
 
 import java.math.BigDecimal;
 import java.util.List;
