+++ src/main/java/kitchenpos/menu/dto/ProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.menu.dto;
 
 import java.math.BigDecimal;
 import kitchenpos.menu.Product;
