+++ src/main/java/kitchenpos/order/dto/OrderLineItemRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.order.dto;
 
 public class OrderLineItemRequest {
