+++ src/main/java/kitchenpos/order/dto/OrderLineItemResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.response;
+package kitchenpos.order.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
