+++ src/main/java/kitchenpos/order/dto/OrderRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.order.dto;
 
 import java.util.List;
