+++ src/main/java/kitchenpos/ordertable/dto/OrderTableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.ordertable.dto;
 
 import kitchenpos.ordertable.OrderTable;
