+++ src/main/java/kitchenpos/ordertable/dto/OrderTableResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.response;
+package kitchenpos.ordertable.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
