+++ src/test/java/kitchenpos/acceptance/MenuAcceptanceTest.java
@@ -3,10 +3,10 @@ package kitchenpos.acceptance;
 import io.restassured.response.ValidatableResponse;
 import java.util.Collections;
 import java.util.List;
-import kitchenpos.dto.request.MenuRequest;
 import kitchenpos.menu.MenuGroup;
 import kitchenpos.menu.MenuProduct;
 import kitchenpos.menu.Product;
+import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.support.RequestBuilder;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.DisplayName;
