+++ src/test/java/kitchenpos/acceptance/TableAcceptanceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.acceptance;
 
 import io.restassured.response.ValidatableResponse;
-import kitchenpos.dto.request.OrderTableRequest;
 import kitchenpos.ordertable.OrderTable;
+import kitchenpos.ordertable.dto.OrderTableRequest;
 import kitchenpos.support.RequestBuilder;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.DisplayName;
